Added table interface and wrapping S4 class
This is in seriously alpha shape but appears
to be working. Most of the features are
in place, though the R interface is rudimentary.

The C++ documentation for module functions
does not get set properly by devtools.
I put some generated Roxygen comments
in cppDoc.R to make it easy to get
the documentation for the two exposed
classes. But devtools mangled it severely,
so another approach is needed, likely
a custom Rd file.

The most basic tests succeed, but more
testing is required.