Eliminated Rdoc comments in modules, new RcppExports.cpp created
For the moment, devtools::document() does not seem to handle
module documentation in C++ files, leading to an R exports
file with NULL's that causes devtools::check() to fail.
Changed //' comments to // for module components until
this changes.

A new file RcppExports.cpp was created then during check(),
which might be enough to solve the import problem for
remote package install.